requirements.txt
Flask==2.3.2
PyJWT==2.8.0
Werkzeug==2.3.6
qdrant-client==2.7.0
langchain==0.1.0
langchain-google-genai==0.0.7
langchain-community==0.0.10
google-generativeai==0.3.0
rank-bm25==0.2.2
pypdf==3.17.1
python-dotenv==1.0.0
